requirements.txt
fastapi
uvicorn
python-dotenv
upsonic
browser-use
playwright
requests
